+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Desktop: Allow more than one media file to be imported from web
+- Printing: Add template variables for pictures
 ---
 * Always add new entries at the very top of this file above other existing entries and this note.
 * Use this layout for new entries: `[Area]: [Details about the change] [reference thread / issue]`